+++ src/components/Experience.ts
@@ -1,7 +1,6 @@
 import { GeneralComponent } from "gamestartr";
-
 import { FullScreenPokemon } from "../FullScreenPokemon";
-import { IPokemon } from "./Battles";
+import { IBattleInfo, IPokemon } from "./Battles";
 
 /**
  * Calculates experience gains and level ups for Pokemon.
@@ -15,7 +14,6 @@ export class Experience<TGameStartr extends FullScreenPokemon> extends GeneralCo
     public levelup(pokemon: IPokemon): void {
         pokemon.level += 1;
         pokemon.statistics = this.gameStarter.equations.newPokemonStatistics(pokemon.title, pokemon.level, pokemon.ev, pokemon.iv);
-
         // TODO: display text box containing levelup info
 
         const evolvedForm: string[] | undefined = this.gameStarter.evolution.checkEvolutions(pokemon);
@@ -38,5 +36,31 @@ export class Experience<TGameStartr extends FullScreenPokemon> extends GeneralCo
             return true;
         }
         return false;
+    }
+
+     /**
+      * Processes experience gain for one or more pokemon
+      *
+      * @param battleInfo   Information about the current battle.
+      * @param partyIsWipedText   Text to be displayed when the opposing pokemon faints.
+      *
+      */
+    public processBattleExperience(battleInfo: IBattleInfo, partyIsWipedText: (string | string[])[][]): void {
+        const experienceToGain =  this.gameStarter.equations.experienceGained(
+            battleInfo.teams.player, battleInfo.teams.opponent);
+
+        // BattleInfo should keep track of all pokemon the player has sent out in the battle and process
+        // them one by one if they have not fainted to allow for shared exp. The same logic will apply
+        // for an exp share where the pokemon holding the exp share is always considered to have been "sent out"
+        partyIsWipedText.push([battleInfo.teams.player.selectedActor.nickname,
+                               " gained ", experienceToGain.toString(), " experience!"]);
+        const levelUp = this.gainExperience(battleInfo.teams.player.selectedActor, experienceToGain);
+        console.log(battleInfo.teams.player.selectedActor.experience);
+        if (levelUp) {
+            partyIsWipedText.push([battleInfo.teams.player.selectedActor.nickname, " grew to level ",
+                                   (battleInfo.teams.player.selectedActor.level).toString(), "!"]);
+        }
+        // Call Learn Move Function which checks level and checks if move can be learned.
+        // LearnMove(Pokemon);
     }
 }
