+++ src/components/battles/animations/shared/Fainting.ts
@@ -46,16 +46,7 @@ export class Fainting<TGameStartr extends FullScreenPokemon> extends GeneralComp
                 const partyIsWipedText: (string | string[])[][] = [[pokemon.nickname, " fainted!"]];
 
                 if (teamName === "opponent") {
-                    const experienceToGain =  this.gameStarter.equations.experienceGained(
-                        battleInfo.teams.player, battleInfo.teams.opponent);
-                    partyIsWipedText.push([battleInfo.teams.player.selectedActor.nickname,
-                                           " gained ", experienceToGain.toString(), " experience!"]);
-                    const levelUp = this.gameStarter.experience.gainExperience
-                    (battleInfo.teams.player.selectedActor, experienceToGain);
-                    if (levelUp) {
-                    partyIsWipedText.push([battleInfo.teams.player.selectedActor.nickname,
-                                           " grew to level ", (battleInfo.teams.player.selectedActor.level).toString(), "!"]);
-                    }
+                    this.gameStarter.experience.processBattleExperience(battleInfo, partyIsWipedText);
                 }
 
                 this.gameStarter.physics.killNormal(thing);
